+++ cjs/src/modules/did.ts
@@ -39,6 +39,7 @@ import { assert } from '@cosmjs/utils-cjs';
 import { PageRequest } from '@cheqd/ts-proto-cjs/cosmos/base/query/v1beta1/pagination';
 import { CheqdQuerier } from '../querier';
 import { DIDDocumentMetadata } from 'did-resolver-cjs';
+import { denormalizeService, normalizeAuthentication, normalizeController, normalizeService } from '../utils';
 
 export const defaultDidExtensionKey = 'did' as const;
 
@@ -548,13 +549,7 @@ export class DIDModule extends AbstractCheqdSDKModule {
 			}
 		});
 
-		const protoService = didDocument?.service?.map((s) => {
-			return Service.fromPartial({
-				id: s?.id,
-				serviceType: s?.type,
-				serviceEndpoint: <string[]>s?.serviceEndpoint,
-			});
-		});
+		const protoService = normalizeService(didDocument);
 
 		return {
 			valid: true,
@@ -598,16 +593,7 @@ export class DIDModule extends AbstractCheqdSDKModule {
 			}
 		});
 
-		const service = protobufDidDocument.service.map((s) => {
-			if (s.serviceType === ServiceType.LinkedDomains)
-				protobufDidDocument.context = [...protobufDidDocument.context, contexts.LinkedDomainsContext];
-
-			return {
-				id: s.id,
-				type: s.serviceType,
-				serviceEndpoint: s.serviceEndpoint,
-			};
-		});
+		const service = denormalizeService(protobufDidDocument);
 
 		const context = (function () {
 			if (protobufDidDocument.context.includes(contexts.W3CDIDv1)) return protobufDidDocument.context;
@@ -685,14 +671,16 @@ export class DIDModule extends AbstractCheqdSDKModule {
 		)
 			return { valid: false, error: 'authentication is required' };
 
+		const normalizedAuthentication = normalizeAuthentication(didDocument);
+
 		// define unique authentication
-		const uniqueAuthentication = new Set<string>(didDocument.authentication as string[]);
+		const uniqueAuthentication = new Set<string>(normalizedAuthentication);
 
 		// validate authentication - case: authentication contains duplicates
-		if (uniqueAuthentication.size < didDocument.authentication!.length)
+		if (uniqueAuthentication.size < normalizedAuthentication.length)
 			return {
 				valid: false,
-				error: `authentication contains duplicate key references: duplicate key reference ${Array.from(uniqueAuthentication).find((a) => didDocument.authentication!.filter((aa) => aa === a).length > 1)}`,
+				error: `authentication contains duplicate key references: duplicate key reference ${Array.from(uniqueAuthentication).find((a) => normalizeAuthentication(didDocument).filter((aa) => aa === a).length > 1)}`,
 			};
 
 		// define unique signatures - shallow, only verificationMethodId, no signature
@@ -713,7 +701,7 @@ export class DIDModule extends AbstractCheqdSDKModule {
 			};
 
 		// define whether external controller or not
-		const externalController = (didDocument.controller as string[]).some((c) => c !== didDocument.id);
+		const externalController = normalizeController(didDocument);
 
 		// validate authentication - case: authentication matches signatures, unique, if no external controller
 		if (!Array.from(uniqueAuthentication).every((a) => uniqueSignatures.has(a)) && !externalController)
@@ -736,7 +724,7 @@ export class DIDModule extends AbstractCheqdSDKModule {
 		if (!querier) throw new Error('querier is required for external controller validation');
 
 		// get external controllers
-		const externalControllers = (didDocument.controller as string[])?.filter((c) => c !== didDocument.id);
+		const externalControllers = externalController.filter((c) => c !== didDocument.id);
 
 		// get external controllers' documents
 		const externalControllersDocuments = await Promise.all(
@@ -761,10 +749,9 @@ export class DIDModule extends AbstractCheqdSDKModule {
 		);
 
 		// define unique required signatures
-		const uniqueRequiredSignatures = new Set([
-			...(didDocument.authentication as string[]),
-			...externalControllersDocuments!.flatMap((d) => d!.authentication as string[]),
-		]);
+		const uniqueRequiredSignatures: Set<string> = new Set(
+			externalControllersDocuments.concat(didDocument).flatMap((d) => (d ? normalizeAuthentication(d) : []))
+		);
 
 		// validate authentication - case: authentication matches signatures, unique, if external controller
 		if (!Array.from(uniqueRequiredSignatures).every((a) => uniqueSignatures.has(a)))
@@ -802,13 +789,14 @@ export class DIDModule extends AbstractCheqdSDKModule {
 			return { valid: false, error: 'authentication is required' };
 
 		// define unique authentication
-		const uniqueAuthentication = new Set<string>(didDocument.authentication as string[]);
+		const authentication = normalizeAuthentication(didDocument);
+		const uniqueAuthentication = new Set(authentication);
 
 		// validate authentication - case: authentication contains duplicates
-		if (uniqueAuthentication.size < didDocument.authentication!.length)
+		if (uniqueAuthentication.size < authentication.length)
 			return {
 				valid: false,
-				error: `authentication contains duplicate key references: duplicate key reference ${Array.from(uniqueAuthentication).find((a) => didDocument.authentication!.filter((aa) => aa === a).length > 1)}`,
+				error: `authentication contains duplicate key references: duplicate key reference ${Array.from(uniqueAuthentication).find((a) => normalizeAuthentication(didDocument).filter((aa) => aa === a).length > 1)}`,
 			};
 
 		// define unique signatures
@@ -832,10 +820,11 @@ export class DIDModule extends AbstractCheqdSDKModule {
 			previousDidDocument = await DIDModule.toSpecCompliantPayload(previousDocument.didDoc);
 		}
 
+		const controllers = normalizeController(didDocument);
+		const previousControllers = normalizeController(previousDidDocument);
+
 		// define whether external controller or not
-		const externalController = (didDocument.controller as string[])
-			.concat(previousDidDocument.controller as string[])
-			.some((c) => c !== didDocument.id);
+		const externalController = controllers.concat(previousControllers).some((c) => c !== didDocument.id);
 
 		// define whether key rotation or not, of any short
 		const keyRotation = !!didDocument.verificationMethod?.some((vm) =>
@@ -850,25 +839,18 @@ export class DIDModule extends AbstractCheqdSDKModule {
 
 		// define controller rotation
 		const controllerRotation =
-			!(didDocument.controller as string[]).every((c) =>
-				(previousDidDocument.controller as string[]).includes(c)
-			) ||
-			!(previousDidDocument.controller as string[]).every((c) =>
-				(didDocument.controller as string[]).includes(c)
-			);
+			!controllers.every((c) => previousControllers.includes(c)) ||
+			!previousControllers.every((c) => controllers.includes(c));
 
 		// define rotated controllers
 		const rotatedControllers = controllerRotation
-			? (previousDidDocument.controller as string[]).filter(
-					(c) => !(didDocument.controller as string[]).includes(c)
-				)
+			? previousControllers.filter((c) => !controllers.includes(c))
 			: [];
 
+		const previousAuthentication = normalizeAuthentication(previousDidDocument);
+
 		// define unique union of authentication
-		const uniqueUnionAuthentication = new Set<string>([
-			...uniqueAuthentication,
-			...(previousDidDocument.authentication as string[]),
-		]);
+		const uniqueUnionAuthentication = new Set<string>([...uniqueAuthentication, ...previousAuthentication]);
 
 		// validate authentication - case: authentication matches signatures, unique, if no external controller, no key rotation
 		if (
@@ -896,10 +878,8 @@ export class DIDModule extends AbstractCheqdSDKModule {
 
 		// define unique union of signatures required, delimited
 		const uniqueUnionSignaturesRequired = new Set([
-			...(didDocument.authentication as string[])
-				.filter((a) => rotatedKeys?.find((rk) => a === rk.id))
-				.map((a) => `${a}(document0)`),
-			...(previousDidDocument.authentication as string[]).map((a) => `${a}(document1)`),
+			...authentication.filter((a) => rotatedKeys?.find((rk) => a === rk.id)).map((a) => `${a}(document0)`),
+			...previousAuthentication.map((a) => `${a}(document1)`),
 		]);
 
 		// define frequency of unique union of signatures required
@@ -963,9 +943,7 @@ export class DIDModule extends AbstractCheqdSDKModule {
 		if (!querier) throw new Error('querier is required for external controller validation');
 
 		// get external controllers
-		const externalControllers = (didDocument.controller as string[])
-			?.filter((c) => c !== didDocument.id)
-			.concat(rotatedControllers);
+		const externalControllers = controllers?.filter((c) => c !== didDocument.id).concat(rotatedControllers);
 
 		// get external controllers' documents
 		const externalControllersDocuments = await Promise.all(
@@ -992,11 +970,11 @@ export class DIDModule extends AbstractCheqdSDKModule {
 		// define unique required signatures, delimited, with external controllers
 		const uniqueUnionSignaturesRequiredWithExternalControllers = new Set<string>([
 			...uniqueUnionSignaturesRequired,
-			...externalControllersDocuments!
-				.flatMap((d) => d!.authentication as string[])
+			...externalControllersDocuments
+				.flatMap((d) => normalizeAuthentication(d))
 				.map(
 					(a) =>
-						`${a}(document${externalControllersDocuments!.findIndex((d) => d?.authentication?.includes(a))})`
+						`${a}(document${externalControllersDocuments.findIndex((d) => normalizeAuthentication(d).includes(a))})`
 				),
 		]);