+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## [5.3.0](https://github.com/cheqd/sdk/compare/5.2.2...5.3.0) (2025-05-27)
+
+### Features
+
+* Add additional fields in service section to support didcomm [DEV-5014] ([#437](https://github.com/cheqd/sdk/issues/437)) ([76a4367](https://github.com/cheqd/sdk/commit/76a43677ec46b167b2f4d64ec5ae2dbf68d85a41))
+* Immutable retrial policy (with enhanced error literals + transaction hash guarantees for asynchronous lookup) ([f67204f](https://github.com/cheqd/sdk/commit/f67204f6f330ef87870b8723f54928e07268009e))
+
+### Bug Fixes
+
+* Normalize controllers and authentication ([#435](https://github.com/cheqd/sdk/issues/435)) ([fe36409](https://github.com/cheqd/sdk/commit/fe364099f961cdbd1f81a06e09b0a7f3b1930fbf))
+
+## [5.3.0-develop.3](https://github.com/cheqd/sdk/compare/5.3.0-develop.2...5.3.0-develop.3) (2025-05-27)
+
+### Features
+
+* Add additional fields in service section to support didcomm [DEV-5014] ([#437](https://github.com/cheqd/sdk/issues/437)) ([76a4367](https://github.com/cheqd/sdk/commit/76a43677ec46b167b2f4d64ec5ae2dbf68d85a41))
+
+## [5.3.0-develop.2](https://github.com/cheqd/sdk/compare/5.3.0-develop.1...5.3.0-develop.2) (2025-05-26)
+
+### Bug Fixes
+
+* Normalize controllers and authentication ([#435](https://github.com/cheqd/sdk/issues/435)) ([fe36409](https://github.com/cheqd/sdk/commit/fe364099f961cdbd1f81a06e09b0a7f3b1930fbf))
+
+## [5.3.0-develop.1](https://github.com/cheqd/sdk/compare/5.2.2...5.3.0-develop.1) (2025-05-22)
+
+### Features
+
+* Immutable retrial policy (with enhanced error literals + transaction hash guarantees for asynchronous lookup) ([f67204f](https://github.com/cheqd/sdk/commit/f67204f6f330ef87870b8723f54928e07268009e))
+
 ## [5.2.2](https://github.com/cheqd/sdk/compare/5.2.1...5.2.2) (2025-05-09)
 
 ### Bug Fixes